move extension functions to extensions package
