+++ app/src/main/java/com/github/libretube/helpers/WindowHelper.kt
@@ -6,6 +6,8 @@ import android.view.WindowManager
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
+import com.github.libretube.ui.extensions.hideSystemBars
+import com.github.libretube.ui.extensions.showSystemBars
 
 object WindowHelper {
     fun toggleFullscreen(activity: Activity, isFullscreen: Boolean) {
@@ -25,19 +27,5 @@ object WindowHelper {
         } else {
             activity.showSystemBars()
         }
-    }
-}
-
-fun Activity.hideSystemBars() {
-    WindowCompat.getInsetsController(window, window.decorView).apply {
-        systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
-        hide(WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.navigationBars())
-    }
-}
-
-fun Activity.showSystemBars() {
-    WindowCompat.getInsetsController(window, window.decorView).apply {
-        systemBarsBehavior = WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
-        show(WindowInsetsCompat.Type.systemBars() or WindowInsetsCompat.Type.navigationBars())
     }
 }
